requirements.txt
# Experiment 1: Pure Exploration Workflow Agent
# Python 3.8+ required

# Anthropic SDK for Claude API
anthropic>=0.40.0

# Environment variable management
python-dotenv>=1.0.0

# Streamlit for web UI
streamlit>=1.30.0

# Qdrant vector database client
qdrant-client>=1.7.0

# OpenAI SDK for embeddings
openai>=1.0.0

# FastAPI for REST API server
fastapi>=0.100.0
uvicorn>=0.23.0
